moved update fixtures
