+++ apps/asap-server/test/handlers/teams/update.test.ts
@@ -6,8 +6,13 @@ import { handler } from '../../../src/handlers/teams/update';
 import { buildGraphQLQueryFetchTeam } from '../../../src/controllers/teams';
 import { apiGatewayEvent } from '../../helpers/events';
 import { identity } from '../../helpers/squidex';
-import * as fixtures from './update.fixtures';
 import decodeToken from '../../../src/utils/validate-token';
+import {
+  getUpdateTeamResponse,
+  updateResponseTeam,
+  getGraphQlTeamResponse,
+  updateExpectation,
+} from '../../fixtures/teams.fixtures';
 
 jest.mock('../../../src/utils/validate-token');
 
@@ -142,16 +147,16 @@ describe('PATCH /teams/{id}', () => {
       .patch(`/api/content/${config.appName}/teams/team-id-1`, {
         tools: { iv: [] },
       })
-      .reply(200, fixtures.getUpdateTeamResponse()) // response is not used
+      .reply(200, getUpdateTeamResponse()) // response is not used
       .post(`/api/content/${config.appName}/graphql`, {
         query: buildGraphQLQueryFetchTeam('team-id-1'),
       })
-      .reply(200, fixtures.getGraphQlTeamResponse())
+      .reply(200, getGraphQlTeamResponse())
       .get(`/api/content/${config.appName}/users`)
       .query({
         $filter: "data/teams/iv/id eq 'team-id-1'",
       })
-      .reply(200, fixtures.usersResponseTeam1);
+      .reply(200, updateResponseTeam);
 
     const result = (await handler(
       apiGatewayEvent({
@@ -169,7 +174,7 @@ describe('PATCH /teams/{id}', () => {
 
     const body = JSON.parse(result.body);
     expect(result.statusCode).toStrictEqual(200);
-    expect(body).toStrictEqual(fixtures.expectation);
+    expect(body).toStrictEqual(updateExpectation);
   });
 
   test('returns 200 when team exists - deletes field', async () => {
@@ -184,16 +189,16 @@ describe('PATCH /teams/{id}', () => {
       .patch(`/api/content/${config.appName}/teams/team-id-1`, {
         tools: { iv: tools },
       })
-      .reply(200, fixtures.getUpdateTeamResponse(tools)) // response is not used
+      .reply(200, getUpdateTeamResponse(tools)) // response is not used
       .post(`/api/content/${config.appName}/graphql`, {
         query: buildGraphQLQueryFetchTeam('team-id-1'),
       })
-      .reply(200, fixtures.getGraphQlTeamResponse(tools))
+      .reply(200, getGraphQlTeamResponse(tools))
       .get(`/api/content/${config.appName}/users`)
       .query({
         $filter: "data/teams/iv/id eq 'team-id-1'",
       })
-      .reply(200, fixtures.usersResponseTeam1);
+      .reply(200, updateResponseTeam);
 
     const result = (await handler(
       apiGatewayEvent({
@@ -217,7 +222,7 @@ describe('PATCH /teams/{id}', () => {
 
     const body = JSON.parse(result.body);
     expect(result.statusCode).toStrictEqual(200);
-    expect(body).toStrictEqual({ ...fixtures.expectation, tools });
+    expect(body).toStrictEqual({ ...updateExpectation, tools });
   });
 
   test('returns 200 when team exists - updates', async () => {
@@ -232,16 +237,16 @@ describe('PATCH /teams/{id}', () => {
       .patch(`/api/content/${config.appName}/teams/team-id-1`, {
         tools: { iv: tools },
       })
-      .reply(200, fixtures.getUpdateTeamResponse(tools)) // response is not used
+      .reply(200, getUpdateTeamResponse(tools)) // response is not used
       .post(`/api/content/${config.appName}/graphql`, {
         query: buildGraphQLQueryFetchTeam('team-id-1'),
       })
-      .reply(200, fixtures.getGraphQlTeamResponse(tools))
+      .reply(200, getGraphQlTeamResponse(tools))
       .get(`/api/content/${config.appName}/users`)
       .query({
         $filter: "data/teams/iv/id eq 'team-id-1'",
       })
-      .reply(200, fixtures.usersResponseTeam1);
+      .reply(200, updateResponseTeam);
 
     const result = (await handler(
       apiGatewayEvent({
@@ -257,6 +262,6 @@ describe('PATCH /teams/{id}', () => {
 
     const body = JSON.parse(result.body);
     expect(result.statusCode).toStrictEqual(200);
-    expect(body).toStrictEqual({ ...fixtures.expectation, tools });
+    expect(body).toStrictEqual({ ...updateExpectation, tools });
   });
 });
